backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
bcrypt==4.0.1
beanie==1.23.6
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
dataclasses-json==0.6.3
distro==1.8.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
fastapi-mail==1.4.1
flake8==7.0.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.345
langchain-core==0.0.9
langsmith==0.0.69
lazy-model==0.2.0
MarkupSafe==2.1.4
marshmallow==3.20.1
mccabe==0.7.0
mongoengine==0.27.0
motor==3.3.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.7
packaging==23.2
passlib==1.7.4
pluggy==1.4.0
pyasn1==0.5.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyflakes==3.2.0
PyJWT==2.8.0
pymongo==4.6.0
pytest==8.0.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==8.2.3
toml==0.10.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.3
